+++ CHANGELOG.md
@@ -8,6 +8,7 @@ CHANGELOG
  * Add support for `SOURCE_DATE_EPOCH` environment variable
  * Add `ResponseEvent::getControllerAttributes()`
  * Add `Request` attribute `_controller_attributes` to decouple controller attributes from their source code
+ * Pass `request` and `args` variables to `Cache` attribute expressions containing the `Request` object and controller arguments
 
 8.0
 ---